Update : 진우 poo 치우기

+++ src/main/java/tools/project/StGuideBook/service/TipPostService.java
@@ -7,6 +7,7 @@ import tools.project.StGuideBook.domain.SiteUser;
 import tools.project.StGuideBook.dto.CommentDTO;
 import tools.project.StGuideBook.dto.TipPostDTO;
 import tools.project.StGuideBook.exception.DataNotFoundException;
+import tools.project.StGuideBook.exception.UnauthorizedException;
 import tools.project.StGuideBook.repository.TipPostRepository;
 
 import java.time.LocalDateTime;
@@ -46,9 +47,14 @@ public class TipPostService {
         return new TipPostDTO(tipPost.getSubject(), tipPost.getContent(), tipPost.getCreateDate(), commentDTOList);
     }
 
-    public TipPost updatePost(Integer author_id, TipPostDTO tipPostDTO) {
-        TipPost tipPost = tipPostRepository.findById(author_id)
+    public TipPost updatePost(Integer id, TipPostDTO tipPostDTO, SiteUser user) {
+        TipPost tipPost = tipPostRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("게시글을 찾지 못했습니다."));
+
+        // 작성자 확인
+        if (!tipPost.getAuthor().equals(user)) {
+            throw new UnauthorizedException("수정 권한이 없습니다.");
+        }
 
         // 수정할 내용 업데이트
         tipPost.setSubject(tipPostDTO.getSubject());
@@ -57,11 +63,15 @@ public class TipPostService {
         return tipPostRepository.save(tipPost);
     }
 
-    public void deletePost(Integer author_id) {
-        TipPost tipPost = tipPostRepository.findById(author_id)
+    public void deletePost(Integer id, SiteUser user) {
+        TipPost tipPost = tipPostRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("게시글을 찾지 못했습니다."));
+
+        // 작성자 확인
+        if (!tipPost.getAuthor().equals(user)) {
+            throw new UnauthorizedException("삭제 권한이 없습니다.");
+        }
 
         tipPostRepository.delete(tipPost);
     }
-
 }
